Properly typing game setup tagged union, cleaned up game setup serialization, moved game models to game_manager module
